Bugfix - Update methods to get numbers to match new headers
V8 headers have been updated, update the code accordingly.
One must pass the current context to functions like:
- `Int32Value`
- `NumberValue`

and it returns a `Maybe<T>` object, here we are sure it matches the
right type as we have checked that just above, so we are safe to call
the `FromJust` method.